// packages/rapiq/src/constants.d.ts
export declare enum Parameter {
    FILTERS = "filters",
    FIELDS = "fields",
    PAGINATION = "pagination",
    RELATIONS = "relations",
    SORT = "sort"
}
export declare enum URLParameter {
    FILTERS = "filter",
    FIELDS = "fields",
    PAGINATION = "page",
    RELATIONS = "include",
    SORT = "sort"
}
export declare const DEFAULT_ID = "__DEFAULT__";
//# sourceMappingURL=constants.d.ts.map